+++ Projet S2/svgfile.h
@@ -21,7 +21,7 @@ class Svgfile
                           double thickness, std::string colorStroke);
         void addTriangle(double x1, double y1, double x2, double y2,
                          double x3, double y3, std::string colorFill=defcol);
-        void addLine(double x1, double y1, double x2, double y2, std::string color=defcol);
+        void addLine(int x1, int y1, int x2, int y2, std::string color=defcol);
         void addArete (int ind1, int ind2, std::string color=defcol);
         void addCross(double x, double y, double span, std::string color=defcol);
